Validate key and value types
We want to restrict the dictionary's key and value types as follows:

Key:
  - needs to be serializable by Unity.
  - cannot be a type derived (directly or indirectly) from UnityEngine.Object.
Because Unity serializes them as references to other objects in the scene or
in the Project (prefabs, scriptable objects, etc.). This means that the object can
be null (and so, not a valid dictionary key).
  - needs to be equatable. So basically, the key can either be a primitive type,
an enum, a struct, or a class that implements Equals and GetHashCode.
This is needed so that the keys are distinguished from one another by value
and not by refence.

Value:
   - needs to be serializable by Unity.